fix: delete user working
