+++ src/constants/endpoints.ts
@@ -12,4 +12,5 @@ export enum Endpoints {
   GET_ALL_TIME_PERIODS = "/time_periods/all",
   CREATE_TIME_PERIOD = "/admin/add_time_period",
   CHANGE_USER_ROLE = "/admin/change_role/{username}",
+  DELETE_USER = "/admin/delete_user/{user_id}",
 }
