+++ src/services/AdminService.ts
@@ -38,3 +38,26 @@ export async function changeUserRole(username: string, roleId: number) {
 
     throw new Error("Failed to change user role");
 }
+
+
+export async function deleteUserById(userId: number) {
+    const response = await HttpClient.delete(
+        formatEndpoint(Endpoints.DELETE_USER, {
+            user_id: userId.toString(),
+        }),
+        true,
+    );
+
+    if (response.ok) {
+        return response.json();
+    }
+
+    switch (response.status) {
+        case 401:
+            throw new Error("Unauthorized");
+        case 404:
+            throw new Error("User not found");
+        default:
+            throw new Error("Failed to delete user");
+    }
+}
