+++ src/services/User.service.ts
@@ -113,25 +113,3 @@ export async function roleNameFromId(roleId: number): Promise<string> {
       throw new Error("Invalid role ID");
   }
 }
-
-export async function deleteUserById(userId: number) {
-  const response = await HttpClient.delete(
-    formatEndpoint(Endpoints.GET_USER_BY_ID, {
-      user_id: userId.toString(),
-    }),
-    true,
-  );
-
-  if (response.ok) {
-    return response.json();
-  }
-
-  switch (response.status) {
-    case 401:
-      throw new Error("Unauthorized");
-    case 404:
-      throw new Error("User not found");
-    default:
-      throw new Error("Failed to delete user");
-  }
-}
